requirements.txt
google-genai
sarvamai

fastapi
uvicorn
gunicorn

python-multipart
python-dotenv
